- Added folder .vs (inner Visual Studio folder) to gitignore list; - Fixed the type of PublicTrade.Timestamp from int to long. Timestamp contains microseconds since 1970.01.01. Now, typical values of datetime (2019.08.26) are 1566821243978 microseconds, it is more int.MaxValue. Now, if you try to fill the Timestamp property with an actual value, an invalid value exception will be thrown.

+++ csharp/src/Org.OpenAPITools/Model/PublicTrade.cs
@@ -212,7 +212,7 @@ namespace Org.OpenAPITools.Model
         /// </summary>
         /// <value>The timestamp of the trade</value>
         [DataMember(Name="timestamp", EmitDefaultValue=false)]
-        public int? Timestamp { get; set; }
+        public long? Timestamp { get; set; }
 
         /// <summary>
         /// The price of the trade
